headers: add static inline to force inlining also in debug
This solves error on compilation fails when compiled with -g.
See https://stackoverflow.com/questions/16245521/c99-inline-function-in-c-file/16245669#16245669.

Basically: inline is treated as optimization, not as replace like preprocessor. So in -O3 inlining is executed but in -g inlining is not executed by default. Since the function is only one line, there will not be problems on debugging.