error handling: replace perror() by fprintf()
Since we are using the app as a CLI, it is weird to display errors and then display success (since non-system errors have errno=0)

Solution 1 is to include errno.h and change errno, but we want to code in a non-posix, so only few errno are common to posix in windows for example. Besides the errno codes are not useful in case the cli is called inside the app via ffi calls, in that case is more useful to provide REST like codes.

Solution 2 is to use fprinf(), it is uglier than perror but we can send custom codes, custom strings etc. We lose the system error messages but we could add them if needed.